requirements.txt
bcrypt==4.3.0
fastapi[standard]==0.115.12
psycopg2-binary==2.9.10
pydantic==2.11.3
pyjwt==2.10.1
python-multipart==0.0.20
sqlalchemy==2.0.40